refactor(data_converter): split source file
