+++ cpp_tools/src/autoware_diffusion_planner_tools/CMakeLists.txt
@@ -21,8 +21,11 @@ endif()
 
 ament_auto_add_executable(data_converter
   src/data_converter.cpp
+  src/io.cpp
+  src/sequence_builder.cpp
+  src/frame_processor.cpp
 )
-target_include_directories(data_converter PRIVATE include)
+target_include_directories(data_converter PRIVATE include src)
 target_link_libraries(data_converter yaml-cpp)
 
 ament_auto_add_executable(inference_tool
